same as previous commit; fix wrong type casting resulting in unsatisfiable conditions

+++ www/history.php
@@ -96,10 +96,10 @@ final class history
 
 		if (($daysago / 365) >= 1) {
 			$daysago = str_replace('.0', '', number_format($daysago / 365, 1));
-			$daysago .= ' Year'.((int) $daysago > 1 ? 's' : '').' Ago';
+			$daysago .= ' Year'.((float) $daysago > 1 ? 's' : '').' Ago';
 		} elseif (($daysago / 30.42) >= 1) {
 			$daysago = str_replace('.0', '', number_format($daysago / 30.42, 1));
-			$daysago .= ' Month'.((int) $daysago > 1 ? 's' : '').' Ago';
+			$daysago .= ' Month'.((float) $daysago > 1 ? 's' : '').' Ago';
 		} elseif ($daysago > 1) {
 			$daysago .= ' Days Ago';
 		} elseif ($daysago == 1) {
